requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
attrs==24.2.0
discord.py==2.4.0
frozenlist==1.4.1
idna==3.10
multidict==6.1.0
propcache==0.2.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
tabulate==0.9.0
yarl==1.15.5
